+++ varnish-sys/src/wrapper.h
@@ -1,3 +1,4 @@
 #include "cache/cache.h"
+#include "cache/cache_filter.h"
 #include "vmod_abi.h"
 #include "vsb.h"